Added authorization logic

+++ controllers/feeds.js
@@ -1,6 +1,7 @@
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1;
@@ -40,19 +41,33 @@ exports.createPost = (req, res, next) => {
     
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
 
     const newPost = new Post({ //creating a mongoose object which gives access to all the mongoose methods
         title: title,
-        content: content
+        content: content,
+        creator: req.userId
     })
 
     newPost.save()
         .then(result => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            creator = user;
+            user.posts.push(newPost);
+            return user.save();
+        })
+        .then(result => {
             res.status(201).json({
                 message: 'Post created successfully',
-                post: result
+                post: newPost,
+                creator: {
+                    _id: creator._id,
+                    name: creator.name
+                }
             })
-        })
+        })    
         .catch( err => {
             if(!err.statusCode) {
                 err.statusCode = 500
@@ -105,6 +120,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if(post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorised');
+                error.statusCode = 403;
+                throw error;
+            }
             post.title = title;
             post.content = content;
 
@@ -133,7 +153,20 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+
+            if(post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorised');
+                error.statusCode = 403;
+                throw error;
+            }
             return Post.findByIdAndRemove(postId);
+        })
+        .then(result => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
         })
         .then(result => {
             console.log(result);
